Add support to see if there are visible notifications in ClockPlugin
to create more lively and intelligent clocks as in Android S

+++ packages/SystemUI/src/com/android/systemui/statusbar/phone/NotificationPanelViewController.java
@@ -894,7 +894,7 @@ public class NotificationPanelViewController extends PanelViewController {
         } else {
             int totalHeight = mView.getHeight();
             int bottomPadding = Math.max(mIndicationBottomPadding, mAmbientIndicationBottomPadding);
-            int clockPreferredY = mKeyguardStatusView.getClockPreferredY(totalHeight);
+            int clockPreferredY = (int) mKeyguardStatusView.getClockPreferredY(totalHeight);
             boolean bypassEnabled = mKeyguardBypassController.getBypassEnabled();
             final boolean
                     hasVisibleNotifications =
